+++ src/layouts/Table/components/TableCell/index.ts
@@ -1,0 +1,2 @@
+export * from "./TableCell.component";
+export type * from "./TableCell.types";