+++ src/layouts/Table/components/index.ts
@@ -1,0 +1,6 @@
+export * from "./TableHeader";
+export * from "./TableHeaderCell";
+export * from "./TableBody";
+export * from "./TableRow";
+export * from "./TableCell";
+export * from "./TableScrollView";