+++ src/layouts/Table/index.ts
@@ -1,0 +1,3 @@
+export * from "./Table.layout";
+export type * from "./Table.types";
+export * from "./components";