+++ src/layouts/index.ts
@@ -2,3 +2,4 @@ export * from "./Flex";
 export * from "./Grid";
 export * from "./GridItem";
 export * from "./Popover";
+export * from "./Table";